Move client extension receive functions to their respective files
