+++ tls/extensions/s2n_client_max_frag_len.c
@@ -32,3 +32,20 @@ int s2n_extensions_client_max_frag_len_send(struct s2n_connection *conn, struct 
 
     return 0;
 }
+
+int s2n_recv_client_max_frag_len(struct s2n_connection *conn, struct s2n_stuffer *extension)
+{
+    if (!conn->config->accept_mfl) {
+        return 0;
+    }
+
+    uint8_t mfl_code;
+    GUARD(s2n_stuffer_read_uint8(extension, &mfl_code));
+    if (mfl_code > S2N_TLS_MAX_FRAG_LEN_4096 || mfl_code_to_length[mfl_code] > S2N_TLS_MAXIMUM_FRAGMENT_LENGTH) {
+        return 0;
+    }
+
+    conn->mfl_code = mfl_code;
+    conn->max_outgoing_fragment_length = mfl_code_to_length[mfl_code];
+    return 0;
+}
