+++ tls/extensions/s2n_client_renegotiation_info.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -16,23 +16,18 @@
 #include <sys/param.h>
 #include <stdint.h>
 
-#include "tls/extensions/s2n_client_sct_list.h"
+#include "tls/extensions/s2n_client_renegotiation_info.h"
 #include "tls/s2n_tls.h"
-#include "tls/s2n_tls_parameters.h"
 
 #include "utils/s2n_safety.h"
 
-int s2n_extensions_client_sct_list_send(struct s2n_connection *conn, struct s2n_stuffer *out)
+int s2n_recv_client_renegotiation_info(struct s2n_connection *conn, struct s2n_stuffer *extension)
 {
-    GUARD(s2n_stuffer_write_uint16(out, TLS_EXTENSION_SCT_LIST));
-    GUARD(s2n_stuffer_write_uint16(out, 0));
+    /* RFC5746 Section 3.2: The renegotiated_connection field is of zero length for the initial handshake. */
+    uint8_t renegotiated_connection_len;
+    GUARD(s2n_stuffer_read_uint8(extension, &renegotiated_connection_len));
+    S2N_ERROR_IF(s2n_stuffer_data_available(extension) || renegotiated_connection_len, S2N_ERR_NON_EMPTY_RENEGOTIATION_INFO);
 
-    return 0;
-}
-
-int s2n_recv_client_sct_list(struct s2n_connection *conn, struct s2n_stuffer *extension)
-{
-    conn->ct_level_requested = S2N_CT_SUPPORT_REQUEST;
-    /* Skip reading the extension, per RFC6962 (3.1.1) it SHOULD be empty anyway  */
+    conn->secure_renegotiation = 1;
     return 0;
 }
