+++ tls/extensions/s2n_client_server_name.c
@@ -44,3 +44,45 @@ int s2n_extensions_client_server_name_send(struct s2n_connection *conn, struct s
 
     return 0;
 }
+
+int s2n_parse_client_hello_server_name(struct s2n_connection *conn, struct s2n_stuffer *extension)
+{
+    if (conn->server_name[0]) {
+        /* already parsed server name extension, exit early */
+        return 0;
+    }
+
+    uint16_t size_of_all;
+    uint8_t server_name_type;
+    uint16_t server_name_len;
+    uint8_t *server_name;
+
+    GUARD(s2n_stuffer_read_uint16(extension, &size_of_all));
+    if (size_of_all > s2n_stuffer_data_available(extension) || size_of_all < 3) {
+        /* the size of all server names is incorrect, ignore the extension */
+        return 0;
+    }
+
+    GUARD(s2n_stuffer_read_uint8(extension, &server_name_type));
+    if (server_name_type != 0) {
+        /* unknown server name type, ignore the extension */
+        return 0;
+    }
+
+    GUARD(s2n_stuffer_read_uint16(extension, &server_name_len));
+    if (server_name_len + 3 > size_of_all) {
+        /* the server name length is incorrect, ignore the extension */
+        return 0;
+    }
+
+    if (server_name_len > sizeof(conn->server_name) - 1) {
+        /* the server name is too long, ignore the extension */
+        return 0;
+    }
+
+    notnull_check(server_name = s2n_stuffer_raw_read(extension, server_name_len));
+
+    /* copy the first server name */
+    memcpy_check(conn->server_name, server_name, server_name_len);
+    return 0;
+}
