+++ tls/extensions/s2n_client_server_name.h
@@ -19,3 +19,4 @@
 #include "stuffer/s2n_stuffer.h"
 
 extern int s2n_extensions_client_server_name_send(struct s2n_connection *conn, struct s2n_stuffer *out);
+extern int s2n_parse_client_hello_server_name(struct s2n_connection *conn, struct s2n_stuffer *extension);
