+++ tls/extensions/s2n_client_session_ticket.c
@@ -19,6 +19,7 @@
 #include "tls/extensions/s2n_client_session_ticket.h"
 #include "tls/s2n_tls.h"
 #include "tls/s2n_tls_parameters.h"
+#include "tls/s2n_resume.h"
 
 #include "utils/s2n_safety.h"
 
@@ -29,6 +30,28 @@ int s2n_extensions_client_session_ticket_send(struct s2n_connection *conn, struc
     GUARD(s2n_stuffer_write_uint16(out, TLS_EXTENSION_SESSION_TICKET));
     GUARD(s2n_stuffer_write_uint16(out, client_ticket_len));
     GUARD(s2n_stuffer_write(out, &conn->client_ticket));
+
+    return 0;
+}
+
+int s2n_recv_client_session_ticket_ext(struct s2n_connection *conn, struct s2n_stuffer *extension)
+{
+    if (conn->config->use_tickets != 1) {
+        /* Ignore the extension. */
+        return 0;
+    }
+
+    /* s2n server does not support session ticket with CLIENT_AUTH enabled */
+    if (s2n_connection_is_client_auth_enabled(conn) > 0) {
+        return 0;
+    }
+
+    if (s2n_stuffer_data_available(extension) == S2N_TICKET_SIZE_IN_BYTES) {
+        conn->session_ticket_status = S2N_DECRYPT_TICKET;
+        GUARD(s2n_stuffer_copy(extension, &conn->client_ticket_to_decrypt, S2N_TICKET_SIZE_IN_BYTES));
+    } else if (s2n_config_is_encrypt_decrypt_key_available(conn->config) == 1) {
+        conn->session_ticket_status = S2N_NEW_TICKET;
+    }
 
     return 0;
 }
