+++ tls/extensions/s2n_client_signature_algorithms.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -15,16 +15,8 @@
 
 #pragma once
 
-#include <stdint.h>
-#include <string.h>
+#include "tls/s2n_connection.h"
+#include "stuffer/s2n_stuffer.h"
 
-#include "utils/s2n_blob.h"
-
-struct s2n_client_hello_parsed_extension {
-	uint16_t extension_type;
-	struct s2n_blob extension;
-};
-
-extern int s2n_client_hello_get_parsed_extension(struct s2n_array *parsed_extensions, s2n_tls_extension_type extension_type,
-        struct s2n_client_hello_parsed_extension *parsed_extension);
-extern void s2n_register_extension(uint16_t ext_type);
+extern int s2n_send_client_signature_algorithms_extension(struct s2n_connection *conn, struct s2n_stuffer *out);
+extern int s2n_recv_client_signature_algorithms(struct s2n_connection *conn, struct s2n_stuffer *extension);
